+++ src/main/java/com/codecool/geometry/shapes/Rectangle.java
@@ -4,7 +4,7 @@ public class Rectangle extends Shape {
     private final float a;
     private final float b;
 
-    public Rectangle(float a, float b, int id) {
+    public Rectangle(int a, int b, int id) {
         super(id);
         this.a = a;
         this.b = b;
